remove support for marshalling core types into custom binary format
