+++ change.go
@@ -2,10 +2,6 @@ package osm
 
 import (
 	"encoding/xml"
-
-	"github.com/paulmach/osm/internal/osmpb"
-
-	"google.golang.org/protobuf/proto"
 )
 
 // Change is the structure of a changeset to be
@@ -63,61 +59,6 @@ func (c *Change) HistoryDatasource() *HistoryDatasource {
 	ds.add(c.Delete, false)
 
 	return ds
-}
-
-// Marshal encodes the osm change data using protocol buffers.
-func (c *Change) Marshal() ([]byte, error) {
-	ss := &stringSet{}
-	encoded := marshalChange(c, ss, true)
-	encoded.Strings = ss.Strings()
-
-	return proto.Marshal(encoded)
-}
-
-// UnmarshalChange will unmarshal the data into a Change object.
-func UnmarshalChange(data []byte) (*Change, error) {
-
-	pbf := &osmpb.Change{}
-	err := proto.Unmarshal(data, pbf)
-	if err != nil {
-		return nil, err
-	}
-
-	return unmarshalChange(pbf, pbf.GetStrings(), nil)
-}
-
-func marshalChange(c *Change, ss *stringSet, includeChangeset bool) *osmpb.Change {
-	if c == nil {
-		return nil
-	}
-
-	return &osmpb.Change{
-		Create: marshalOSM(c.Create, ss, includeChangeset),
-		Modify: marshalOSM(c.Modify, ss, includeChangeset),
-		Delete: marshalOSM(c.Delete, ss, includeChangeset),
-	}
-}
-
-func unmarshalChange(encoded *osmpb.Change, ss []string, cs *Changeset) (*Change, error) {
-	var err error
-	c := &Change{}
-
-	c.Create, err = unmarshalOSM(encoded.Create, ss, cs)
-	if err != nil {
-		return nil, err
-	}
-
-	c.Modify, err = unmarshalOSM(encoded.Modify, ss, cs)
-	if err != nil {
-		return nil, err
-	}
-
-	c.Delete, err = unmarshalOSM(encoded.Delete, ss, cs)
-	if err != nil {
-		return nil, err
-	}
-
-	return c, nil
 }
 
 // MarshalXML implements the xml.Marshaller method to allow for the
