--- internal/osmpb/generate.go
@@ -1,3 +1,0 @@
-package osmpb
-
-//go:generate protoc  --go_opt=module=github.com/paulmach/osm/internal/osmpb  --go_out=. osm.proto
