+++ osm.go
@@ -2,13 +2,8 @@ package osm
 
 import (
 	"encoding/xml"
-	"errors"
 	"fmt"
 	"regexp"
-
-	"github.com/paulmach/osm/internal/osmpb"
-
-	"google.golang.org/protobuf/proto"
 )
 
 // These values should be returned if the osm data is actual
@@ -42,16 +37,6 @@ type OSM struct {
 	Changesets Changesets `xml:"changeset"`
 	Notes      Notes      `xml:"note"`
 	Users      Users      `xml:"user"`
-}
-
-// Marshal encodes the osm data using protocol buffers.
-// Will only save the elements: nodes, ways and relations.
-func (o *OSM) Marshal() ([]byte, error) {
-	ss := &stringSet{}
-	encoded := marshalOSM(o, ss, true)
-	encoded.Strings = ss.Strings()
-
-	return proto.Marshal(encoded)
 }
 
 // Append will add the given object to the OSM object.
@@ -196,122 +181,6 @@ func (o *OSM) HistoryDatasource() *HistoryDatasource {
 
 	ds.add(o)
 	return ds
-}
-
-// UnmarshalOSM will unmarshal the data into an OSM object.
-func UnmarshalOSM(data []byte) (*OSM, error) {
-
-	pbf := &osmpb.OSM{}
-	err := proto.Unmarshal(data, pbf)
-	if err != nil {
-		return nil, err
-	}
-
-	return unmarshalOSM(pbf, pbf.GetStrings(), nil)
-}
-
-// includeChangeset can be set to false to not repeat the changeset
-// info every item, if this comes from osm change data.
-func marshalOSM(o *OSM, ss *stringSet, includeChangeset bool) *osmpb.OSM {
-	encoded := &osmpb.OSM{}
-	if o == nil {
-		return nil
-	}
-
-	if len(o.Nodes) > 0 {
-		encoded.DenseNodes = marshalNodes(o.Nodes, ss, includeChangeset)
-	}
-
-	if len(o.Ways) > 0 {
-		encoded.Ways = make([]*osmpb.Way, len(o.Ways))
-		for i, w := range o.Ways {
-			encoded.Ways[i] = marshalWay(w, ss, includeChangeset)
-		}
-	}
-
-	if len(o.Relations) > 0 {
-		encoded.Relations = make([]*osmpb.Relation, len(o.Relations))
-		for i, r := range o.Relations {
-			encoded.Relations[i] = marshalRelation(r, ss, includeChangeset)
-		}
-	}
-
-	if o.Bounds != nil {
-		encoded.Bounds = &osmpb.Bounds{
-			MinLat: geoToPInt64(o.Bounds.MinLat),
-			MaxLat: geoToPInt64(o.Bounds.MaxLat),
-			MinLon: geoToPInt64(o.Bounds.MinLon),
-			MaxLon: geoToPInt64(o.Bounds.MaxLon),
-		}
-	}
-
-	return encoded
-}
-
-func unmarshalOSM(encoded *osmpb.OSM, ss []string, cs *Changeset) (*OSM, error) {
-	if encoded == nil {
-		return nil, nil
-	}
-
-	o := &OSM{}
-	if len(encoded.Nodes) != 0 && encoded.DenseNodes != nil {
-		return nil, errors.New("found both nodes and dense nodes")
-	}
-
-	if len(encoded.Nodes) != 0 {
-		o.Nodes = make([]*Node, len(encoded.Nodes))
-		for i, en := range encoded.Nodes {
-			n, err := unmarshalNode(en, ss, cs)
-			if err != nil {
-				return nil, err
-			}
-
-			o.Nodes[i] = n
-		}
-	}
-
-	if encoded.DenseNodes != nil {
-		var err error
-		o.Nodes, err = unmarshalNodes(encoded.DenseNodes, ss, cs)
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	if len(encoded.Ways) != 0 {
-		o.Ways = make([]*Way, len(encoded.Ways))
-		for i, ew := range encoded.Ways {
-			w, err := unmarshalWay(ew, ss, cs)
-			if err != nil {
-				return nil, err
-			}
-
-			o.Ways[i] = w
-		}
-	}
-
-	if len(encoded.Relations) != 0 {
-		o.Relations = make([]*Relation, len(encoded.Relations))
-		for i, er := range encoded.Relations {
-			r, err := unmarshalRelation(er, ss, cs)
-			if err != nil {
-				return nil, err
-			}
-
-			o.Relations[i] = r
-		}
-	}
-
-	if encoded.Bounds != nil {
-		o.Bounds = &Bounds{
-			MinLat: float64(encoded.Bounds.GetMinLat()) / locMultiple,
-			MaxLat: float64(encoded.Bounds.GetMaxLat()) / locMultiple,
-			MinLon: float64(encoded.Bounds.GetMinLon()) / locMultiple,
-			MaxLon: float64(encoded.Bounds.GetMaxLon()) / locMultiple,
-		}
-	}
-
-	return o, nil
 }
 
 // MarshalJSON allows the tags to be marshalled as an object
