+++ relation_test.go
@@ -232,50 +232,6 @@ func TestRelation_MarshalXML(t *testing.T) {
 	}
 }
 
-func TestRelations_Marshal(t *testing.T) {
-	rs := Relations{
-		{ID: 123},
-		{ID: 321},
-	}
-
-	data, err := rs.Marshal()
-	if err != nil {
-		t.Fatalf("relations marshal error: %v", err)
-	}
-
-	rs2, err := UnmarshalRelations(data)
-	if err != nil {
-		t.Fatalf("relations unmarshal error: %v", err)
-	}
-
-	if !reflect.DeepEqual(rs, rs2) {
-		t.Errorf("relations not equal")
-		t.Logf("%+v", rs)
-		t.Logf("%+v", rs2)
-	}
-
-	// empty relations
-	rs = Relations{}
-
-	data, err = rs.Marshal()
-	if err != nil {
-		t.Fatalf("relations marshal error: %v", err)
-	}
-
-	if l := len(data); l != 0 {
-		t.Errorf("length of relation data should be 0, got %v", l)
-	}
-
-	rs2, err = UnmarshalRelations(data)
-	if err != nil {
-		t.Fatalf("relations unmarshal error: %v", err)
-	}
-
-	if rs2 != nil {
-		t.Errorf("should return nil Relations for empty data, got %v", rs2)
-	}
-}
-
 func TestMember_ids(t *testing.T) {
 	cases := []struct {
 		name string
