+++ update.go
@@ -4,8 +4,6 @@ import (
 	"fmt"
 	"sort"
 	"time"
-
-	"github.com/paulmach/osm/internal/osmpb"
 )
 
 // CommitInfoStart is the start time when we know committed at information.
@@ -84,91 +82,4 @@ func (us updatesSortIndex) Less(i, j int) bool {
 	}
 
 	return us[i].Timestamp.Before(us[j].Timestamp)
-}
-
-func marshalUpdates(updates Updates) *osmpb.DenseMembers {
-	if len(updates) == 0 {
-		return nil
-	}
-
-	l := len(updates)
-	indexes := make([]int32, l)
-	versions := make([]int32, l)
-	timestamps := make([]int64, l)
-	changesetIDs := make([]int64, l)
-	lats := make([]int64, l)
-	lons := make([]int64, l)
-	revs := make([]int32, l)
-
-	hasLoc := false
-	hasRev := false
-	for i, u := range updates {
-		indexes[i] = int32(u.Index)
-		versions[i] = int32(u.Version)
-		timestamps[i] = timeToUnix(u.Timestamp)
-		changesetIDs[i] = int64(u.ChangesetID)
-		if u.Lat != 0 || u.Lon != 0 {
-			hasLoc = true
-			lats[i] = geoToInt64(u.Lat)
-			lons[i] = geoToInt64(u.Lon)
-		}
-
-		if u.Reverse {
-			hasRev = true
-			revs[i] = 1
-		}
-	}
-
-	result := &osmpb.DenseMembers{
-		Indexes:      encodeInt32(indexes),
-		Versions:     versions,
-		ChangesetIds: encodeInt64(changesetIDs),
-		Timestamps:   encodeInt64(timestamps),
-	}
-
-	if hasLoc {
-		result.Lats = encodeInt64(lats)
-		result.Lons = encodeInt64(lons)
-	}
-
-	if hasRev {
-		result.Orientation = revs
-	}
-
-	return result
-}
-
-func unmarshalUpdates(encoded *osmpb.DenseMembers) Updates {
-	if encoded == nil {
-		return nil
-	}
-
-	result := make([]Update, len(encoded.Indexes))
-
-	decodeInt32(encoded.Indexes)
-	decodeInt64(encoded.ChangesetIds)
-	decodeInt64(encoded.Timestamps)
-
-	decodeInt64(encoded.Lats)
-	decodeInt64(encoded.Lons)
-
-	for i := range encoded.Indexes {
-		result[i] = Update{
-			Index:       int(encoded.Indexes[i]),
-			Version:     int(encoded.Versions[i]),
-			ChangesetID: ChangesetID(encoded.ChangesetIds[i]),
-			Timestamp:   unixToTime(encoded.Timestamps[i]),
-		}
-
-		if len(encoded.Lats) > i {
-			result[i].Lat = float64(encoded.Lats[i]) / locMultiple
-			result[i].Lon = float64(encoded.Lons[i]) / locMultiple
-		}
-
-		if len(encoded.Orientation) > i && encoded.Orientation[i] > 0 {
-			result[i].Reverse = true
-		}
-	}
-
-	return result
 }
