+++ way_test.go
@@ -317,50 +317,6 @@ func TestWay_MarshalXML(t *testing.T) {
 	}
 }
 
-func TestWays_Marshal(t *testing.T) {
-	ws := Ways{
-		{ID: 123},
-		{ID: 321},
-	}
-
-	data, err := ws.Marshal()
-	if err != nil {
-		t.Fatalf("ways marshal error: %v", err)
-	}
-
-	ws2, err := UnmarshalWays(data)
-	if err != nil {
-		t.Fatalf("ways unmarshal error: %v", err)
-	}
-
-	if !reflect.DeepEqual(ws, ws2) {
-		t.Errorf("ways not equal")
-		t.Logf("%+v", ws)
-		t.Logf("%+v", ws2)
-	}
-
-	// empty ways
-	ws = Ways{}
-
-	data, err = ws.Marshal()
-	if err != nil {
-		t.Fatalf("ways marshal error: %v", err)
-	}
-
-	if l := len(data); l != 0 {
-		t.Errorf("length of way data should be 0, got %v", l)
-	}
-
-	ws2, err = UnmarshalWays(data)
-	if err != nil {
-		t.Fatalf("ways unmarshal error: %v", err)
-	}
-
-	if ws2 != nil {
-		t.Errorf("should return nil Ways for empty data, got %v", ws2)
-	}
-}
-
 func TestWayNodes_ids(t *testing.T) {
 	wns := WayNodes{
 		{ID: 1, Version: 3},
